Add ReflectionPropertyOps utils and tests
Introduce ReflectionPropertyOps (hpp/cpp) to centralize property access, serialization/deserialization, widget resolution and copy helpers. Update ReflectionSerializer, ReflectionWidgets, Reflector, InspectorPanel to use the new utilities and export ReflectionSerializer with ENGINE_API. Change PropertyDescriptor.offset and memberOffset() to size_t for wider portability and add safe enum/int storage helpers. Add ReflectionTests executable and tests covering registry/widget resolution, component copy, entity serialization round-trip and invalid JSON handling; include the new test in the top-level Tests buildscript.

+++ Engine/src/Reflection/ReflectionTypes.hpp
@@ -78,7 +78,7 @@ struct PropertyDescriptor
 {
     const char* name;           // Field name (e.g. "speed")
     EPropertyType type;         // Data type
-    uint32_t offset;            // offsetof() from component start
+    size_t offset;              // offsetof() from component start
     uint32_t size;              // sizeof the field
     PropertyMeta meta;          // Editor metadata
 };
